Decouple logic for applying mutation to local db and enqueuing worker
